pyproject.toml
[tool.poetry]
name = "tg"
version = "0.1.0"
description = ""
authors = ["caio"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "1.5.3"
happybase = "^1.2.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
